Move integration test to cairo_run_test.go
