+++ pkg/hints/bigint_utils.go
@@ -1,52 +1,15 @@
-package hint_utils
+package hints
 
 import (
 	"math/big"
 
-	"github.com/lambdaclass/cairo-vm.go/pkg/lambdaworks"
+	. "github.com/lambdaclass/cairo-vm.go/pkg/hints/hint_utils"
 	. "github.com/lambdaclass/cairo-vm.go/pkg/lambdaworks"
-	. "github.com/lambdaclass/cairo-vm.go/pkg/types"
 	. "github.com/lambdaclass/cairo-vm.go/pkg/vm"
-	"github.com/lambdaclass/cairo-vm.go/pkg/vm/memory"
 	. "github.com/lambdaclass/cairo-vm.go/pkg/vm/memory"
+
 	"github.com/pkg/errors"
 )
-
-/*
-Implements hint:
-%{
-    from starkware.cairo.common.cairo_secp.secp_utils import split
-
-    segments.write_arg(ids.res.address_, split(value))
-%}
-*/
-
-func NondetBigInt3(virtual_machine VirtualMachine, exec_scopes ExecutionScopes, ids_data IdsManager) error {
-	res_relloc, err := ids_data.GetAddr("res", &virtual_machine)
-	if err != nil {
-		return err
-	}
-
-	value_uncast, err := exec_scopes.Get("value")
-	if err != nil {
-		return err
-	}
-	value := value_uncast.(big.Int)
-
-	bigint3_split, err := Bigint3Split(value)
-	if err != nil {
-		return err
-	}
-	arg := make([]memory.MaybeRelocatable, 0)
-
-	for i := 0; i < 3; i++ {
-		m := memory.NewMaybeRelocatableFelt(lambdaworks.FeltFromBigInt(&bigint3_split[i]))
-		arg = append(arg, *m)
-	}
-
-	virtual_machine.Segments.LoadData(res_relloc, &arg)
-	return nil
-}
 
 // In cairo, various structs are used to represent big integers, all of them have numbered fields of Felt type (d0, d1,...) and share the same behaviours
 // This file contains an implementation of each behaviour at the limbs level, and the wrappers for each specific type
